% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_clean_dataframe.R
\name{reddit_clean_dataframe}
\alias{reddit_clean_dataframe}
\title{reddit_clean_dataframe()
This function returns a cleaned dataframe ready for sentiment analysis}
\arguments{
\item{input_url}{optional input, if you don't want to use the reddit_raw_dataframe function}

\item{dataframe}{a compatible dataframe that just neesd cleaning applied}
}
\value{
a cleaned dataframe, 3 columns
}
\description{
reddit_clean_dataframe()
This function returns a cleaned dataframe ready for sentiment analysis
}
\examples{
reddit_clean_dataframe("https://www.reddit.com/r/funny/comments/7mjw12/my_cab_driver_tonight_was_so_excited_to_share/")
reddit_clean_dataframe(dataframe)
}
